improve UI on the overall form for CRUD forms

+++ ProjectF.WebUI/Pages/Products/Product.cs
@@ -8,10 +8,10 @@ namespace ProjectF.WebUI.Pages.Products
         public string Description { get; set; }
         public string Reference { get; set; }
         public Category Category { get; set; }
-        public long CategoryId { get; set; }
-        public long WerehouseId { get; set; }
+        public int CategoryId { get; set; }
+        public int WerehouseId { get; set; }
         public Warehouse Werehouse { get; set; }
-        public long TaxId { get; set; }
+        public int TaxId { get; set; }
         public Tax Tax { get; set; }
         public bool IsService { get; set; }
         public decimal Cost { get; set; }
